Refactor SourceFromCollection...
