+++ client/src/components/Collections/wizard/SourceFromCollectionApplyRules.vue
